requirements.txt
pytest==8.4.2
allure-pytest==2.15.2
requests==2.32.5
PyYAML==6.0.3
colorlog==6.10.1